+++ package.json
@@ -37,7 +37,6 @@
     "@blitzjs/next": "2.0.9",
     "@blitzjs/rpc": "2.0.9",
     "@prisma/client": "5.4.2",
-    "@tanstack/react-query": "5.44.0",
     "@vercel/analytics": "1.0.1",
     "autoprefixer": "10.4.14",
     "blitz": "2.0.9",
@@ -48,7 +47,7 @@
     "gray-matter": "4.0.3",
     "highlight.js": "11.9.0",
     "next": "14.2.3",
-    "next-themes": "0.2.1",
+    "next-themes": "0.3.0",
     "nextjs-google-analytics": "2.3.3",
     "postcss": "8.4.24",
     "postmark": "4.0.2",
@@ -66,6 +65,7 @@
     "remark-parse": "11.0.0",
     "remark-rehype": "11.1.0",
     "secure-password": "4.0.0",
+    "sharp": "0.33.4",
     "tailwindcss": "3.3.2",
     "unified": "11.0.4",
     "xml-beautifier": "0.5.0",
