+++ src/app/styles/Home.module.css
@@ -101,7 +101,6 @@
   background-position: 100% 100%, 0 100%;
   background-repeat: no-repeat;
   transition: background-size 400ms;
-  text-decoration: underline;
 }
 
 .textLink:hover,
