+++ tailwind.config.js
@@ -1,5 +1,5 @@
-
 module.exports = {
+  darkMode: "class",
   content: ["./{src,app,pages}/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
@@ -69,5 +69,4 @@ module.exports = {
     },
   },
   plugins: [require("@tailwindcss/typography")],
-  darkMode: 'class',
 }
